Added better retry logic for curl, and also for android build which is flaky due to dependency HTTP connections failing.

+++ build_scripts/android/install_prereqs.sh
@@ -15,12 +15,15 @@ elif [[ $(uname) == "Linux" ]]; then
 else
     platform=windows
     # On Windows, we have an additional dependency for Strings
-    curl -LSs \
-	 --connect-timeout 30 \
-	 --retry 5 \
-	 --retry-delay 10 \
-        "https://download.sysinternals.com/files/Strings.zip" \
-         --output Strings.zip
+    set +e
+    for retry in 1 2 3 4 5 error; do
+	if [[ $retry == "error" ]]; then exit 5; fi
+	curl -LSs \
+             "https://download.sysinternals.com/files/Strings.zip" \
+             --output Strings.zip && break
+	sleep 10
+    done
+    set -e
     unzip -q Strings.zip && rm -f Strings.zip
 fi
 
@@ -64,15 +67,15 @@ if [[ -z "${NDK_ROOT}" || -z $(grep "Pkg\.Revision = 16\." "${NDK_ROOT}/source.p
 	    echo "Error, could not run 'curl' to download NDK. Is it in your PATH?"
 	    exit 1
 	fi
-	set -x
-	curl -LSs \
-	     --connect-timeout 30 \
-	     --retry 5 \
-	     --retry-delay 10 \
-	     "https://dl.google.com/android/repository/android-ndk-r16b-${platform}-x86_64.zip" \
-	     --output /tmp/android-ndk-r16b.zip
+	set +e
+	for retry in 1 2 3 4 5 error; do
+	    if [[ $retry == "error" ]]; then exit 5; fi
+	    curl -LSs \
+		 "https://dl.google.com/android/repository/android-ndk-r16b-${platform}-x86_64.zip" \
+		 --output /tmp/android-ndk-r16b.zip && break
+	    sleep 10
+	set -e
 	(cd /tmp && unzip -q android-ndk-r16b.zip && rm -f android-ndk-r16b.zip)
-	set +x
 	echo "NDK r16b has been downloaded into /tmp/android-ndk-r16b"
     fi
 fi
